backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Python backend for Optiq"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4==4.12.3",
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
    "pipecat-ai[deepgram,daily,openai,silero,webrtc]>=0.0.76",
    "pypdf==4.3.1",
    "python-dotenv==1.0.1",
    "tiktoken==0.7.0",
]

[tool.uv]
override-dependencies = [
    "numba==0.61.2",
]
